tests: Split tests into separate files
